+++ internal/core/items/text.rs
@@ -1005,6 +1005,10 @@ impl TextInput {
         );
     }
 
+    pub fn clear_selection(self: Pin<&Self>, _: &Rc<dyn WindowAdapter>, _: &ItemRc) {
+        self.as_ref().anchor_position_byte_offset.set(self.as_ref().cursor_position_byte_offset());
+    }
+
     fn select_word(self: Pin<&Self>, window_adapter: &Rc<dyn WindowAdapter>, self_rc: &ItemRc) {
         let text = self.text();
         let anchor = self.anchor_position(&text);
@@ -1233,6 +1237,19 @@ pub unsafe extern "C" fn slint_textinput_select_all(
     let window_adapter = &*(window_adapter as *const Rc<dyn WindowAdapter>);
     let self_rc = ItemRc::new(self_component.clone(), self_index);
     Pin::new_unchecked(&*text_input).as_ref().select_all(window_adapter, &self_rc);
+}
+
+#[cfg(feature = "ffi")]
+#[no_mangle]
+pub unsafe extern "C" fn slint_textinput_clear_selection(
+    text_input: *const TextInput,
+    window_adapter: *const crate::window::ffi::WindowAdapterRcOpaque,
+    self_component: &vtable::VRc<crate::component::ComponentVTable>,
+    self_index: usize,
+) {
+    let window_adapter = &*(window_adapter as *const Rc<dyn WindowAdapter>);
+    let self_rc = ItemRc::new(self_component.clone(), self_index);
+    Pin::new_unchecked(&*text_input).as_ref().clear_selection(window_adapter, &self_rc);
 }
 
 #[cfg(feature = "ffi")]