+++ internal/interpreter/eval.rs
@@ -625,6 +625,7 @@ fn call_builtin_function(
                 {
                     match &*name {
                         "select-all" => textinput.select_all(&window_adapter, &item_rc),
+                        "clear-selection" => textinput.clear_selection(&window_adapter, &item_rc),
                         "cut" => textinput.cut(&window_adapter, &item_rc),
                         "copy" => textinput.copy(&window_adapter, &item_rc),
                         "paste" => textinput.paste(&window_adapter, &item_rc),