+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project are documented in this file.
 
 ### Slint Language
 
+- Added `clear-selection()` to `TextInput`, `LineEdit`, and `TextEdit`.
+
 ### Rust
 
 ### C++